refactor: change folder structure & rename file
