+++ app/3. widgets/panel/ui/panel-navbar/index.tsx
