+++ app/6. shared/api/ens/model.ts
