+++ app/6. shared/helpers/encoder-decoder/base.ts
