+++ app/6. shared/helpers/encoder-decoder/index.ts
