+++ app/6. shared/helpers/local-storage/index.ts
@@ -1,4 +1,4 @@
-import LocalStorageManager from "@shared/helpers/localStorage/base";
+import LocalStorageManager from "@shared/helpers/local-storage/base";
 
 const localStorage = new LocalStorageManager();
 export default localStorage;
