+++ app/6. shared/helpers/local-storage/model.ts
