+++ app/6. shared/helpers/local-storage/types.ts
