+++ app/6. shared/service/model.ts
